using cli backup instead of bb repo, misc bugfixes, gcal iframe stuff, refactoring welcome message to only fire when door has been opened+motion trigger+wifi change, etc

+++ automation/jsr223/00_scripts/01_mqtt.js
@@ -206,7 +206,7 @@ JSRule({
 			var lat  = json.lat;
             var lon  = json.lon;
             var acc  = json.acc;
-            var batt = parseInt(json.batt);
+            var batt = parseFloat(json.batt);
 
             var geocodeURL = "https://maps.googleapis.com/maps/api/geocode/json?latlng="+lat+","+lon+"&key=XXXXXXX"
             var geocodeJson = JSON.parse(HTTP.sendHttpGetRequest(geocodeURL));
